moved template-create-class
