+++ src/processing/blossoms/files/template_files/template_create_blossom.cpp
